Added bd murphy logo and authorbdmurphy.com as default url.
Updated to use gradle plugin 8.13.2

Version to 0.5.1
Updated to have text imbedded into images and show the wifi password.

Added new logo.
Change to version 0.6.0 to release.

Ready for release - maybe. First time trying.
build for closed testing.

BUild for open testing.

Fix the problem with long strings getting trucated.

+++ app/src/main/java/com/bdqrgen/util/QRCodeGenerator.kt
@@ -75,7 +75,7 @@ object QRCodeGenerator {
             textAlign = Paint.Align.LEFT
         }
         
-        val maxWidth = QR_SIZE - 40
+        val maxWidth = (QR_SIZE - 40).toFloat()
         val wrappedLines = wrapText(text, paint, maxWidth)
         
         val lineHeight = paint.fontMetrics.bottom - paint.fontMetrics.top + 8
